backend/requirements.txt
asgiref==3.8.1
black==24.4.0
click==8.1.7
distlib==0.3.8
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
filelock==3.13.4
flake8==7.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
sqlparse==0.5.0
virtualenv==20.25.1
